Article component is defined out of page
